requirements.txt
uvicorn
fastapi
python-dotenv
requests
streamlit